+++ library/src/main/java/org/mustangproject/ZUGFeRD/DXExporterFromA3.java
@@ -64,6 +64,8 @@ import org.apache.xmpbox.xml.XmpParsingException;
 import org.apache.xmpbox.xml.XmpSerializer;
 import org.mustangproject.EStandard;
 import org.mustangproject.FileAttachment;
+import static org.mustangproject.util.StringUtils.isBlank;
+import static org.mustangproject.util.StringUtils.isNotBlank;
 
 import jakarta.activation.DataSource;
 import jakarta.activation.FileDataSource;
@@ -573,7 +575,7 @@ public class DXExporterFromA3 extends ZUGFeRDExporterFromA3 {
 	@Override
   protected void writeAdobePDFSchema(XMPMetadata xmp) {
 		AdobePDFSchema pdf = getAdobePDFSchema(xmp);
-		if (overwrite || isEmpty(pdf.getProducer()))
+		if (overwrite || isBlank(pdf.getProducer()))
 			pdf.setProducer(producer);
 	}
 
@@ -595,7 +597,7 @@ public class DXExporterFromA3 extends ZUGFeRDExporterFromA3 {
 	@Override
   protected void writePDFAIdentificationSchema(XMPMetadata xmp) {
 		PDFAIdentificationSchema pdfaid = getPDFAIdentificationSchema(xmp);
-		if (overwrite || isEmpty(pdfaid.getConformance())) {
+		if (overwrite || isBlank(pdfaid.getConformance())) {
 			try {
 				pdfaid.setConformance(conformanceLevel.getLetter());
 			} catch (BadFieldValueException ex) {
@@ -631,14 +633,14 @@ public class DXExporterFromA3 extends ZUGFeRDExporterFromA3 {
 
 		ArrayProperty titleProperty = dc.getTitleProperty();
 		if (titleProperty != null) {
-			if (overwrite && !isEmpty(title)) {
+			if (overwrite && isNotBlank(title)) {
 				dc.removeProperty(titleProperty);
 				dc.setTitle(title);
 			} else if (titleProperty.getElementsAsString().stream().anyMatch("Untitled"::equalsIgnoreCase)) {
 				// remove unfitting ghostscript default
 				dc.removeProperty(titleProperty);
 			}
-		} else if (!isEmpty(title)) {
+		} else if (isNotBlank(title)) {
 			dc.setTitle(title);
 		}
 	}
@@ -657,7 +659,7 @@ public class DXExporterFromA3 extends ZUGFeRDExporterFromA3 {
 	@Override
   protected void writeXMLBasicSchema(XMPMetadata xmp) {
 		XMPBasicSchema xsb = getXmpBasicSchema(xmp);
-		if (overwrite || isEmpty(xsb.getCreatorTool()) || "UnknownApplication".equals(xsb.getCreatorTool()))
+		if (overwrite || isBlank(xsb.getCreatorTool()) || "UnknownApplication".equals(xsb.getCreatorTool()))
 			xsb.setCreatorTool(creatorTool);
 		if (overwrite || xsb.getCreateDate() == null)
 			xsb.setCreateDate(Calendar.getInstance());
@@ -683,15 +685,15 @@ public class DXExporterFromA3 extends ZUGFeRDExporterFromA3 {
 			info.setCreationDate(Calendar.getInstance());
 		if (overwrite || info.getModificationDate() == null)
 			info.setModificationDate(Calendar.getInstance());
-		if (overwrite || (isEmpty(info.getAuthor()) && !isEmpty(author)))
+		if (overwrite || (isBlank(info.getAuthor()) && isNotBlank(author)))
 			info.setAuthor(author);
-		if (overwrite || (isEmpty(info.getProducer()) && !isEmpty(fullProducer)))
+		if (overwrite || (isBlank(info.getProducer()) && isNotBlank(fullProducer)))
 			info.setProducer(fullProducer);
-		if (overwrite || (isEmpty(info.getCreator()) && !isEmpty(creator)))
+		if (overwrite || (isBlank(info.getCreator()) && isNotBlank(creator)))
 			info.setCreator(creator);
-		if (overwrite || (isEmpty(info.getTitle()) && !isEmpty(title)))
+		if (overwrite || (isBlank(info.getTitle()) && isNotBlank(title)))
 			info.setTitle(title);
-		if (overwrite || (isEmpty(info.getSubject()) && !isEmpty(subject)))
+		if (overwrite || (isBlank(info.getSubject()) && isNotBlank(subject)))
 			info.setSubject(subject);
 	}
 
@@ -772,15 +774,5 @@ public class DXExporterFromA3 extends ZUGFeRDExporterFromA3 {
 		DAPullProvider z2p = new DAPullProvider();
 		setXMLProvider(z2p);
 		return this;
-	}
-
-	/**
-	 * Utility method inspired by apache commons-lang3 StringUtils.
-	 *
-	 * @param string the string to test
-	 * @return true if the string is null or empty
-	 */
-	private boolean isEmpty(String string) {
-		return string == null || string.isEmpty();
 	}
 }