+++ History.md
@@ -1,7 +1,17 @@
+2.24.x
+=======
+- #1150 Remove [D].[M].[Y] after Rechnungsdatum in HTML visualization
+- #996  Omit ExemptionReason written to line item level for profile EN16931, some validators raise a warning or even an error
+- #1097 Invoice importer: LegalOrganization id not imported (wrong node name 'GlobalID' instead of 'ID' used)
+- #1151 'BR-FXEXT-IC-08rev' in ZF 2.5 requires VAT exemption reason text and code for allowances and charges
+- #1145 Document-level allowance with VAT category O emits RateApplicablePercent (violates BR-O-06)
+
+
 2.24.0
 =======
 2026-06-11
 
+- #1143 LineID may not be written empty in DeliveryNoteReferencedDocument
 - #1141 ZUGFeRD 2.5/Factur-X 1.09
 - #1127'Fix possible StringIndexOutOfBoundsException when extracting XMP.'
 - #1133 incorrect warning in Subinvoiceline calculation